This is an automated PR to promote changes from `dev` to `master`.
Please review and test before merging.

## For Testers
See [TESTING.md](./TESTING.md) for complete testing instructions.

## Suggested Merge Schedule
According to our release policy, this PR is expected to be merged on:
**Friday**
Testers are encouraged to test the changes before merging.
Please note that this schedule may be adjusted based on the needs of the
project.
## Commits to be merged:
---
* chore breaking(input): Unset `# force_no_accel` (ca5f3169) by kRHYME7
* fix [Bug]: Importing themes #614 (188890e8) by kRHYME7
* fixup: manage `hyprsunset` through `hyprctl` to avoid color flickering
(#613) (547afe02) by Ilya Bogdanov
* fixup: improve launch of system monitor if it was closed without
`sysmonlaunch.sh` (#616) (c3199587) by Ilya Bogdanov
* fixup: improve padding of `arch` glyph (#615) (cdfc615f) by Ilya
Bogdanov
* chore Fixup rofi launch config (#611) (16c5ff02) by Ilya Bogdanov
* chore refactor: use hyprlock provided `$LAYOUT` for detecting current
keyboard layout (#609) (6acd79fd) by Ilya Bogdanov
* fix(zsh): zsh-autosuggestions won't work on first prompt when deferred
(#605) (843855a9) by Khing
---

Please review the changes carefully before merging.

+++ Configs/.local/lib/hyde/sysmonlaunch.sh
@@ -48,12 +48,15 @@ pidFile="$HYDE_RUNTIME_DIR/sysmonlaunch.pid"
 if [ -f "$pidFile" ]; then
   while IFS= read -r line; do
     pid=$(awk -F ':::' '{print $1}' <<<"$line")
-    cmd=$(awk -F ':::' '{print $2}' <<<"$line")
-    pkill -P "$pid"
-    pkg_installed flatpak && flatpak kill "$cmd" 2>/dev/null
+    if [ -d "/proc/${pid}" ]; then
+      cmd=$(awk -F ':::' '{print $2}' <<<"$line")
+      pkill -P "$pid"
+      pkg_installed flatpak && flatpak kill "$cmd" 2>/dev/null
+      rm "$pidFile"
+      exit 0
+    fi
   done <"$pidFile"
   rm "$pidFile"
-  exit 0
 fi
 
 pkgChk=("io.missioncenter.MissionCenter" "htop" "btop" "top")                     # Array of commands to check
